refactor(dashboard): extract planning action chips
Move the plan, vBRIEF, and tasks chip behavior into shared components.
Keep IssueCard bead gating intact while FeatureCard reuses the same chip logic.
